Can't release because of goreleaser limitation Temporal fix
