+++ cmd/nem/main.go
@@ -1,7 +1,8 @@
 // Copyright 2017 The nem-toolchain project authors. All rights reserved.
 // Use of this source code is governed by a MIT license that can be found in the LICENSE file.
 
-package nem
+// Command nem responses for command line user interface
+package main
 
 import (
 	"bufio"
@@ -18,6 +19,75 @@ import (
 	"github.com/nem-toolchain/nem-toolchain/pkg/vanity"
 	"github.com/urfave/cli"
 )
+
+var (
+	// date stores build timestamp
+	date string
+	// commit stores actual commit hash
+	commit string
+	// version indicates actual version
+	version string
+)
+
+func main() {
+	app := cli.NewApp()
+	app.Name = "nem"
+	app.Usage = "command-line toolchain for NEM blockchain"
+
+	if version == "" {
+		app.Version = "git"
+	} else {
+		app.Version = fmt.Sprintf("%v (%v / %v)", version, commit, date)
+	}
+
+	app.Flags = []cli.Flag{
+		cli.StringFlag{
+			Name:   "chain",
+			Value:  "mainnet",
+			EnvVar: "NEM_CHAIN,CHAIN",
+			Usage:  "chain id from `CHAIN`: [mainnet|mijin|testnet]",
+		},
+	}
+
+	app.Commands = []cli.Command{
+		accountCommand(),
+	}
+
+	_ = app.Run(os.Args)
+}
+
+func chainGlobalOption(c *cli.Context) (core.Chain, error) {
+	var ch core.Chain
+	switch c.GlobalString("chain") {
+	case "mijin":
+		ch = core.Mijin
+	case "mainnet":
+		ch = core.Mainnet
+	case "testnet":
+		ch = core.Testnet
+	default:
+		return ch, fmt.Errorf("unknown chain '%v'", c.GlobalString("chain"))
+	}
+	return ch, nil
+}
+
+//
+//
+//import (
+//"bufio"
+//"encoding/hex"
+//"fmt"
+//"math"
+//"os"
+//"runtime"
+//"strings"
+//"time"
+//
+//"github.com/nem-toolchain/nem-toolchain/pkg/core"
+//"github.com/nem-toolchain/nem-toolchain/pkg/keypair"
+//"github.com/nem-toolchain/nem-toolchain/pkg/vanity"
+//"github.com/urfave/cli"
+//)
 
 func accountCommand() cli.Command {
 	return cli.Command{
@@ -106,7 +176,10 @@ func infoAction(c *cli.Context) error {
 	if err != nil {
 		return cli.NewExitError(err.Error(), 1)
 	}
-	if (fi.Mode() & os.ModeNamedPipe) != 0 {
+
+	fmt.Println(">> DEBUG mode: ", (fi.Mode()&os.ModeDevice) != 0)
+
+	if (fi.Mode() & os.ModeDevice) != 0 {
 		fmt.Print("Enter private key: ")
 	}
